requirements.txt
Flask
flask-cors
python-dotenv
PyPDF2
requests
vapi_server_sdk
